Improved subannual plots
When plotting multiple layers in a subannual plot, first filter out for
all layers, any date in which any layer is NA. If we don't do this, we
will be plotting the mean subannual timeseries over a *different* time
period for each different layer, if the layers cover different time
periods.

If the layers cover different time periods (ie no overlap exists), then
we revert to the old behaviour, and plot them anyway, even though this
will make this plot hard to interpret. Not sure what else we could do
here.